fix: 경매 및 낙찰 테이블 수정 및 등록시 상품명,결제기한 추가
summary:

footer:

+++ common/src/main/java/com/dev_high/common/dto/WinningOrderRecommendationResponse.java
@@ -1,10 +1,11 @@
 package com.dev_high.common.dto;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 public record WinningOrderRecommendationResponse(
     String productId,
-    Long winningAmount,
+    BigDecimal winningAmount,
     OffsetDateTime winningDate,
     String status
 ) {
